Address comments - move calendarpicker into its own componenet
