+++ package.json
@@ -23,6 +23,7 @@
     "i18n-iso-countries": "^6.0.0",
     "leaflet": "^1.6.0",
     "lodash": "^4.17.15",
+    "moment": "2.29.1",
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-compound-slider": "^2.5.0",
